Timer works, is obj-oriented
Timer functionality is now contained within a timer object and it seems to work. 
Buttons also now adjust based on the state of the timer. State of the timer uses an enumerator.
Also improved input validation by using an Input class, which allows the default values to be dynamic based on the last approved value.
Still some other things to change, will move to GitHub issues for working on.